Add updateUser in CRUD for users
An updateUser function has been created in the crud-users service. This function is designed for updating user data in the DynamoDB. Additionally, a PUT route to engage this function has been added to the Serverless configuration. This route listens to "users/{id}" path, making it possible to update users based on their id.